Change homework number from 12 to 13, split some components to more granular
