--- homework_12_create-react-app/fish-shop/src/api-mock/api.js
@@ -1,23 +1,0 @@
-export const apiResponse = {
-  status: 'Ok',
-  result: [
-    {
-      name: 'Палтус'
-    },
-    {
-      name: 'Сёмга'
-    },
-    {
-      name: 'Coм'
-    },
-    {
-      name: 'Мойва'
-    },
-    {
-      name: 'Сельдь'
-    },
-    {
-      name: 'Тунец'
-    }
-  ]
-};
