--- homework_12_create-react-app/fish-shop/src/index.css
@@ -1,38 +1,0 @@
-body,
-html {
-  font-family: "Open Sans", "Arial", sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  padding: 0;
-  margin: 0;
-  height: 100%;
-}
-
-h1 {
-  margin: 0;
-}
-@media (max-width: 850px) {
-  h1 {
-    font-size: 21px;
-  }
-}
-
-h2 {
-  margin: 14px 0 0 0;
-  font-size: 32px;
-}
-
-h3 {
-  padding-top: 24px;
-}
-
-p {
-  margin: 0;
-}
-
-.a_theme_light {
-  color: RoyalBlue;
-}
-.a_theme_dark {
-  color: PowderBlue;
-}
