+++ homework_13_create-react-app/fish-shop/src/App.css
@@ -22,16 +22,6 @@
   flex-direction: column;
 }
 
-a, p {
-  padding: 0 12px;
-}
-
-h3, h2, h1 {
-  padding: 0 12px;
-  font-weight: bold;
-  font-size: 24px;
-}
-
 .page-layout {
   height: 100vh;
   grid-template-columns: 240px 1fr auto;
@@ -191,39 +181,5 @@ h3, h2, h1 {
   }
   .product-filter__type-block {
     padding: 0 24px;
-  }
-}
-
-#loader {
-  position: absolute;
-  width: 64px;
-  height: 64px;
-  top: 10%;
-  right: 2%;
-  padding: 15.5px;
-  z-index: 99;
-  background-color: #ffffff;
-  border: 1px solid #d0d7de;
-  border-radius: 32px;
-  box-shadow: 0 8px 24px rgba(140, 149, 159, 0.2);
-  box-sizing: border-box;
-  display: block;
-}
-
-.anim-rotate {
-  box-sizing: content-box;
-  color: #218bff;
-  animation-name: rotate;
-  animation-duration: 1200ms;
-  animation-iteration-count: 10;
-  animation-direction: normal;
-  animation-timing-function: cubic-bezier(0.85, 0.35, 0.13, 0.66);
-}
-@keyframes rotate {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
   }
 }
