+++ homework_13_create-react-app/fish-shop/src/api-mock/api.js
@@ -1,0 +1,35 @@
+export const apiResponse = {
+  status: 'Ok',
+  result: {
+    categories: [
+      {
+        name: 'Замороженные рыбы',
+        description: 'Мы заморозили рыбов для вас'
+      },
+      {
+        name: 'Живые рыбы',
+        description: 'На кухню или на разведение'
+      }
+    ],
+    products: [
+      {
+        name: 'Палтус'
+      },
+      {
+        name: 'Сёмга'
+      },
+      {
+        name: 'Coм'
+      },
+      {
+        name: 'Мойва'
+      },
+      {
+        name: 'Сельдь'
+      },
+      {
+        name: 'Тунец'
+      }
+    ]
+  }
+};
