+++ homework_13_create-react-app/fish-shop/src/components/aside/Aside.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import {ProductFilter} from "../product-filter/ProductFilter";
+
+export class Aside extends React.Component<any, any> {
+  render() {
+    return (
+      <aside className="page-layout__aside">
+        <ProductFilter/>
+      </aside>
+    );
+  }
+}
