+++ homework_13_create-react-app/fish-shop/src/components/product-tile/ProductTile.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import './ProductTile.css';
+import {ProductTileButtons} from "./__buttons/ProductTileButtons";
 
 export class ProductTile extends React.Component<any, any> {
   render() {
@@ -8,11 +9,8 @@ export class ProductTile extends React.Component<any, any> {
         <img
           src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkmAcAAKQAoJgx3+MAAAAASUVORK5CYII="
           alt="Fish" width="60" height="60" />
-        <div className="product-tile__buttons">
-          <a className="a_theme_light" href="/product.html">{this.props.name}</a>
-          <button className="product-tile__button product-tile__button_theme_light">Купить</button>
-        </div>
+        <ProductTileButtons name={this.props.name}/>
       </article>
-    )
+    );
   }
 }
