+++ homework_13_create-react-app/fish-shop/src/forms/showcase/Showcase.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import {Categories} from "../../components/categories/Categories";
+import {Products} from "../../components/products/Products";
+
+export class Showcase extends React.Component<any, any> {
+  render() {
+    return (
+      <main className="page-layout__main page-layout__main_theme_light">
+        <Categories/>
+        <Products/>
+      </main>
+    );
+  }
+}
