+++ homework_13_create-react-app/fish-shop/src/index.css
@@ -1,0 +1,82 @@
+body,
+html {
+  font-family: "Open Sans", "Arial", sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  padding: 0;
+  margin: 0;
+  height: 100%;
+}
+
+a, p {
+  padding: 0 12px;
+}
+
+h3, h2, h1 {
+  padding: 0 12px;
+  font-weight: bold;
+  font-size: 24px;
+}
+
+h1 {
+  margin: 0;
+}
+@media (max-width: 850px) {
+  h1 {
+    font-size: 21px;
+  }
+}
+
+h2 {
+  margin: 14px 0 0 0;
+  font-size: 32px;
+}
+
+h3 {
+  padding-top: 24px;
+}
+
+p {
+  margin: 0;
+}
+
+.a_theme_light {
+  color: RoyalBlue;
+}
+.a_theme_dark {
+  color: PowderBlue;
+}
+
+#loader {
+  position: absolute;
+  width: 64px;
+  height: 64px;
+  top: 10%;
+  right: 2%;
+  padding: 15.5px;
+  z-index: 99;
+  background-color: #ffffff;
+  border: 1px solid #d0d7de;
+  border-radius: 32px;
+  box-shadow: 0 8px 24px rgba(140, 149, 159, 0.2);
+  box-sizing: border-box;
+  display: block;
+}
+
+.anim-rotate {
+  box-sizing: content-box;
+  color: #218bff;
+  animation-name: rotate;
+  animation-duration: 1200ms;
+  animation-iteration-count: 10;
+  animation-direction: normal;
+  animation-timing-function: cubic-bezier(0.85, 0.35, 0.13, 0.66);
+}
+@keyframes rotate {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
